# backend/docs/swagger.yaml
basePath: /api
definitions:
  dto.AIQuestionRequest:
    properties:
      language:
        example: en
        type: string
      previous_response_id:
        example: "12345"
        type: string
      question:
        example: What is the capital of France?
        maxLength: 256
        type: string
    required:
    - question
    type: object
  dto.AIQuestionResponse:
    properties:
      answer:
        example: The capital of France is Paris.
        type: string
      response_id:
        example: "67890"
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      new_password:
        example: StrongP@ssw0rd
        minLength: 8
        type: string
      token:
        example: abcd1234efgh
        type: string
    required:
    - new_password
    - token
    type: object
...
  dto.SendNotificationEmailRequest:
    properties:
      body:
        example: The system will be down for maintenance at midnight.
        type: string
      subject:
        example: System maintenance
        type: string
      to:
        example: admin@example.com
        type: string
    required:
    - body
    - subject
    - to
    type: object
  dto.SendResetPasswordEmailRequest:
    properties:
      language:
        example: en
        type: string
      to:
        example: user@example.com
        type: string
    required:
    - to
    type: object
  dto.SendVerificationEmailRequest:
    properties:
      language:
        example: en
        type: string
      to:
        example: user@example.com
        type: string
    required:
    - to
    type: object
...
  dto.ValidateTokenRequest:
    properties:
      token:
        example: abcd1234efgh
        type: string
      token_type:
        example: reset_password
        type: string
    required:
    - token
    - token_type
    type: object
  dto.VerifyAccountRequest:
    properties:
      token:
        example: verifytoken123
        type: string
    required:
    - token
    type: object
...
info:
  contact:
    email: dev@example.com
    name: API Support
    url: https://example.com/support
  description: Public HTTP API for the Fitness Tracker app.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /ai/ask-general:
    post:
      consumes:
      - application/json
      description: Lets the user ask AI any general-purpose question.
      parameters:
      - description: Question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AIQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Ask a general AI question
      tags:
      - ai
  /ai/ask-stats:
    post:
      consumes:
      - application/json
      description: Lets the user ask AI about stats-related information.
      parameters:
      - description: Question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AIQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Ask a stats-related question
      tags:
      - ai
  /ai/ask-workouts:
    post:
      consumes:
      - application/json
      description: Lets the user ask AI about workouts and training.
      parameters:
      - description: Question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AIQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIQuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Ask a workouts-related question
      tags:
      - ai
...
  /email/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password reset payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Reset password with token
      tags:
      - email
  /email/send-account-verification:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification email payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendVerificationEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Send account verification email
      tags:
      - email
  /email/send-notification:
    post:
      consumes:
      - application/json
      parameters:
      - description: Notification email payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendNotificationEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Send notification email (admin only)
      tags:
      - email
  /email/send-reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset password email payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendResetPasswordEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Send reset password email
      tags:
      - email
  /email/validate-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token validation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ValidateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Invalid/expired token
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Validate email-related token
      tags:
      - email
  /email/verify-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account verification payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Verify account with token
      tags:
      - email
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"